Add a launcher patch for the LAC demo.
This also includes tie-ins with Raptor7 (lac-demo-raptor.pd in the
raptor7:lac-demo branch) so that both patches can be played in concert.
Specifically, the switch-ab subpatch allows quick switching between the
two melody lines on channel 1, by muting channel 1 on one patch and
unmuting the other. This is also bound to CC64 (sustain pedal).