Add a toggle for syncing Raptor to the demo patch.
This takes care of starting raptor playback in sync with algojuggle
without relying on MIDI clock sync. Note that you still have to make
sure that both are running at the same tempo. Also, both algojuggle and
raptor need to be running in the same Pd instance to make this work.